refactor: implement logic by not using getter
